DEVOPS_ETL/DEVOPS_ETL/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devops-etl"
version = "0.1.0"
description = "Extraction et analyse des données depuis les outils DevOps pour générer des KPIs"
authors = [
    {name = "Equipe DevOps ETL"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "openpyxl>=3.1.2",
    "pandas>=2.1.1",
    "dependency-injector>=4.41.0",
    "pydantic>=2.4.2",  # Note: Migration de Pydantic v1 à v2 nécessite des modifications de code
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "black>=23.9.1",
    "isort>=5.12.0",
    "mypy>=1.5.1",
    "flake8>=6.1.0",
    "pylint>=3.0.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
